fix: sonarqube issues and duplications
