+++ seeds/parserInitialGoodOpenApi3.1.json
@@ -9,7 +9,7 @@
     },
     "servers":[
        {
-          "url":"http://petstore.swagger.io/v1"
+          "url":"https://petstore.swagger.io/v1"
        }
     ],
     "paths":{
