+++ src/parser/openapi3/body.js
@@ -4,7 +4,7 @@
 
 const _ = require('lodash')
 
-module.exports = function() {
+module.exports = function () {
 
   const MAX_DEPTH_LEVEL = 20;
   let seenSchemas = new WeakSet();
@@ -12,19 +12,23 @@ module.exports = function() {
   return function get(verb, path, bodyResponse) {
     seenSchemas = new WeakSet();
 
-  	if (!_.isObject(global.definition.paths)) {
-      require('../../utils/error.js')('paths is required')
+    const hasPaths = _.isObject(global.definition.paths);
+    const hasWebhooks = _.isObject(global.definition.webhooks);
+
+    if (!hasPaths) {
+      if (hasWebhooks) return bodyResponse ? {} : undefined;
+      require('../../utils/error.js')('paths is required');
     }
     const endpoint = global.definition.paths[path][_.toLower(verb)]
     if (!endpoint) return undefined;
 
-    if (!bodyResponse){
+    if (!bodyResponse) {
       let body = endpoint['requestBody'];
       if (!body) return undefined;
 
       if (body['$ref']) {
         let componentType = body['$ref'].split('/')[2]
-        const ref = _.replace(body['$ref'], '#/components/'+componentType+'/', '')
+        const ref = _.replace(body['$ref'], '#/components/' + componentType + '/', '')
         body = global.definition.components[componentType][ref]
       }
       if (!body.content['application/json']) {
@@ -32,10 +36,21 @@ module.exports = function() {
         let schema = undefined;
         let properContent = false;
         let index = 0;
-        // Búsqueda de un content type soportado por la aplicación
         while (!properContent && index < contentsArray.length) {
-          const withoutRefs = replaceRefs(body.content[contentsArray[index]].schema, 1);
-          switch(contentsArray[index]) {
+          const mediaType = body.content[contentsArray[index]];
+
+          if (!mediaType.schema) {
+            if (contentsArray[index] === 'application/octet-stream') {
+              return {
+                type: 'string',
+                format: 'binary'
+              };
+            }
+            continue;
+          }
+
+          const withoutRefs = replaceRefs(mediaType.schema, 1);
+          switch (contentsArray[index]) {
             case 'application/x-www-form-urlencoded':
               schema = replaceAllOfs(withoutRefs);
               properContent = true;
@@ -51,6 +66,16 @@ module.exports = function() {
           index++;
         }
 
+        if (!properContent && contentsArray.length > 0) {
+          const firstContentType = contentsArray[0];
+          const mediaType = body.content[firstContentType];
+
+          if (mediaType && mediaType.schema) {
+            const withoutRefs = replaceRefs(mediaType.schema, 1);
+            return replaceAllOfs(withoutRefs);
+          }
+        }
+
         if (!properContent) console.warn('\x1b[33m%s\x1b[0m', `Warning: cannot create body due to unknown content type (Endpoint: ${verb} ${path})`);
         return schema;
       }
@@ -58,7 +83,7 @@ module.exports = function() {
       return replaceAllOfs(withOutRefs);
     }
     const bodyResponses = {}
-    _.forEach(endpoint['responses'], function(response, status) {
+    _.forEach(endpoint['responses'], function (response, status) {
       if (response && response.content && response.content['application/json'] && response.content['application/json'].schema) {
         const withOutRefs = replaceRefs(response.content['application/json'].schema, 1)
         bodyResponses[status] = replaceAllOfs(withOutRefs)
@@ -66,7 +91,7 @@ module.exports = function() {
           const requiredWtihoutDuplicates = bodyResponses[status].required.filter((value, index, arr) => {
             return arr.indexOf(value) === index;
           });
-          bodyResponses[status].required = requiredWtihoutDuplicates; 
+          bodyResponses[status].required = requiredWtihoutDuplicates;
         }
       }
     })
@@ -74,35 +99,40 @@ module.exports = function() {
   }
 
   function replaceRefs(schema, depthLevel) {
+    if (!_.isObject(schema)) {
+      return schema;
+    }
     let result = {}
     if (depthLevel < MAX_DEPTH_LEVEL) {
       for (const i in schema) {
         if (i === '$ref') {
           let componentType = schema[i].split('/')[2]
-          const ref = _.replace(schema[i], '#/components/'+componentType+'/', '')
-          if (checkCircularReferences(ref,3,2) || checkCircularReferences(ref,3,3) || checkCircularReferences(ref,3,4)){
-            return { type: 'string',
-              description: 'Circular REF solved swagger2postman' 
+          const ref = _.replace(schema[i], '#/components/' + componentType + '/', '')
+          if (checkCircularReferences(ref, 3, 2) || checkCircularReferences(ref, 3, 3) || checkCircularReferences(ref, 3, 4)) {
+            return {
+              type: 'string',
+              description: 'Circular REF solved swagger2postman'
             }
           }
 
           let entity = global.definition.components[componentType][ref]
           if (!entity) {
-            require('../../utils/error.js')('ref '+ref+' is not defined')
+            require('../../utils/error.js')('ref ' + ref + ' is not defined')
           }
 
           entity = replaceRefs(entity, depthLevel + 1);
-          result = _.merge(result, entity)
-        } else if ( _.isArray(schema[i]) && i !== 'required') {
+          result = _.merge({}, result, entity)
+        } else if (_.isArray(schema[i]) && i !== 'required') {
           const arrayResult = []
-          if (i === 'example'){
-            continue
+          if (i === 'example' || i === 'examples') {
+            result[i] = schema[i];
+            continue;
           }
           for (const k in schema[i]) {
             arrayResult.push(replaceRefs(schema[i][k], depthLevel + 1))
           }
           result[i] = arrayResult
-        } else if ( _.isObject(schema[i]) && i !== 'required') {
+        } else if (_.isObject(schema[i]) && i !== 'required') {
           result[i] = replaceRefs(schema[i], depthLevel + 1);
         } else {
           result[i] = schema[i];
@@ -112,10 +142,9 @@ module.exports = function() {
     return result
   }
 
-  function replaceAllOfs(schema){
+  function replaceAllOfs(schema) {
     if (!_.isObject(schema)) return schema;
 
-   // Detectar ciclos por identidad
     if (seenSchemas.has(schema)) {
       return {
         type: "string",
@@ -125,70 +154,101 @@ module.exports = function() {
     seenSchemas.add(schema);
 
     let result = {}
+
+    if (Array.isArray(schema.type)) {
+      const nonNullType = schema.type.find(t => t !== 'null');
+      result.type = nonNullType || schema.type[0];
+    }
+
     for (let i in schema) {
-      if (i === 'allOf' && _.isArray(schema[i])){
-        let merged = {'required':[],'properties':{},'type':'object'}
+      if (i === 'type' && Array.isArray(schema.type)) {
+        continue;
+      }
+      if (i === 'allOf' && _.isArray(schema[i])) {
+        let merged = { 'required': [], 'properties': {}, 'type': 'object' }
         for (let t in schema[i]) {
 
           if (schema[i][t]['type'] === 'string') {
             merged = schema[i][t]
           } else {
-          
+
             for (let k in schema[i][t]) {
-              if (k === 'type'){
+              if (k === 'type') {
                 merged['type'] = schema[i][t][k]
-              } else if (k === 'required'){
-                merged['required'] = _.concat(merged['required'],schema[i][t]['required'])
-              } else if (k === 'properties'){
-                for (let z in schema[i][t]['properties']){
+              } else if (k === 'required') {
+                merged['required'] = _.concat(merged['required'], schema[i][t]['required'])
+              } else if (k === 'properties') {
+                for (let z in schema[i][t]['properties']) {
                   merged['properties'][z] = replaceAllOfs(schema[i][t]['properties'][z])
                 }
-              } else if (k === 'allOf'){
+              } else if (k === 'allOf') {
                 let downSchema = replaceAllOfs(schema[k])
                 if (downSchema['0']) {
                   downSchema = downSchema['0']
                 }
-                merged['required'] = _.concat(merged['required'],downSchema['required'])
-                merged['properties'] = _.merge(merged['properties'],downSchema['properties'])
+                merged['required'] = _.concat(merged['required'], downSchema['required'])
+                merged['properties'] = _.merge(merged['properties'], downSchema['properties'])
                 continue
               } else if (k === 'description') {
                 continue
               } else if (k === 'items') {
                 continue
-              }else {
-                console.warn('the property '+k+' of allOf is not implemented')
+              } else {
+                console.warn('the property ' + k + ' of allOf is not implemented')
               }
             }
 
           }
 
         }
-        result = _.merge(result, merged)
-      } else if ( _.isArray(schema[i]) && i !== 'required') {
+        result = _.merge({}, result, merged)
+      } else if (_.isArray(schema[i]) && i !== 'required') {
         if (schema[i].every(v => !_.isObject(v))) {
           result[i] = [...schema[i]];
         } else {
           result[i] = schema[i].map(item => replaceAllOfs(item));
         }
-      } else if ( _.isObject(schema[i]) && i !== 'required') {
-        // result.type = 'object';
-        result[i] = _.merge(result[i],replaceAllOfs(schema[i]))
+      } else if (_.isObject(schema[i]) && i !== 'required') {
+        const value = replaceAllOfs(schema[i]);
+
+        if (_.isPlainObject(value) && _.isPlainObject(result[i])) {
+          result[i] = _.merge({}, result[i], value);
+        } else {
+          result[i] = value;
+        }
       } else {
         result[i] = schema[i]
       }
     }
+    if (typeof result.exclusiveMinimum === 'number') {
+      result.minimum = result.exclusiveMinimum;
+      result.exclusiveMinimum = true;
+    }
+
+    if (typeof result.exclusiveMaximum === 'number') {
+      result.maximum = result.exclusiveMaximum;
+      result.exclusiveMaximum = true;
+    }
+
+    if (schema.contentEncoding) {
+      result.contentEncoding = schema.contentEncoding;
+    }
+
+    if (schema.contentMediaType) {
+      result.contentMediaType = schema.contentMediaType;
+    }
     return result
   }
-  
-  function checkCircularReferences(reference,depthLevel,patternNumber){
-    if (!global.circularTail){
+
+  function checkCircularReferences(reference, depthLevel, patternNumber) {
+    if (!global.circularTail) {
       global.circularTail = []
     }
-    if (! global.circularTail[patternNumber] ){
+    if (!global.circularTail[patternNumber]) {
       global.circularTail[patternNumber] = [reference]
       return false
     }
-    if (global.circularTail[patternNumber].length < (depthLevel * patternNumber) ){
+    if (global.circularTail[patternNumber].length < (depthLevel * patternNumber)) {
       global.circularTail[patternNumber].push(reference)
       return false
     }
@@ -201,15 +261,15 @@ module.exports = function() {
     let lastArray = groups[0]
     comparation:
     for (let i = 1; i < groups.length; i++) {
-      for (let k in groups[i]){
-        if (groups[i][k] !== groups[i - 1][k]){
+      for (let k in groups[i]) {
+        if (groups[i][k] !== groups[i - 1][k]) {
           areEquals = false
           break comparation
         }
       }
     }
 
-    if (areEquals){
+    if (areEquals) {
       global.circularTail = []
     }
