+++ test/parser-body.js
@@ -4,7 +4,7 @@ const assert = require('assert')
 
 describe('parser-body', () => {
 
-  it('void request swagger2', () => {
+  it('body void request swagger2', () => {
 
     global.definition = require('../seeds/parserInitialGood.json')
 
@@ -12,7 +12,7 @@ describe('parser-body', () => {
     assert.deepStrictEqual(body, undefined)
   })
 
-  it('void request openapi3.0', () => {
+  it('body void request openapi3.0', () => {
 
     global.definition = require('../seeds/parserInitialGoodOpenApi3.json')
 
@@ -20,7 +20,7 @@ describe('parser-body', () => {
     assert.deepStrictEqual(body, undefined)
   })
 
-  it('void request openapi3.0 2', () => {
+  it('body void request openapi3.0 2', () => {
 
     global.definition = require('../seeds/parserInitialGoodOpenApi3.json')
 
@@ -28,15 +28,15 @@ describe('parser-body', () => {
     assert.deepStrictEqual(body, undefined)
   })
 
-  it('void request openapi3.1', () => {
+  it('body void request openapi3.1', () => {
 
     global.definition = require('../seeds/parserInitialGoodOpenApi3.1.json')
 
-    const body = require('../src/parser/openapi31/body.js')('GET','/pets')
+    const body = require('../src/parser/openapi3/body.js')('GET','/pets')
     assert.deepStrictEqual(body, undefined)
   })
 
-  it('fill request swagger2', () => {
+  it('body fill request swagger2', () => {
 
     global.definition = require('../seeds/parserInitialGood.json')
 
@@ -44,7 +44,7 @@ describe('parser-body', () => {
     assert.deepStrictEqual(body, {"type":"object","required":["name"],"properties":{"name":{"type":"string"},"tag":{"type":"string"}}})
   })
 
-  it('fill request openapi3.0', () => {
+  it('body fill request openapi3.0', () => {
 
     global.definition = require('../seeds/parserInitialGoodOpenApiExpanded3.json')
 
@@ -52,15 +52,15 @@ describe('parser-body', () => {
     assert.deepStrictEqual(body, {"type":"object","required":["name"],"properties":{"name":{"type":"string"},"tag":{"type":"string"}}})
   })
 
-  it('fill request openapi3.1', () => {
+  it('body fill request openapi3.1', () => {
 
     global.definition = require('../seeds/parserInitialGoodOpenApiExpanded3.1.json')
 
-    const body = require('../src/parser/openapi31/body.js')('POST','/pets')
+    const body = require('../src/parser/openapi3/body.js')('POST','/pets')
     assert.deepStrictEqual(body, {"type":"object","required":["name"],"properties":{"name":{"type":"string"},"tag":{"type":"string"}}})
   })
 
-  it('items, ref and allof response swagger2', () => {
+  it('body items, ref and allof response swagger2', () => {
 
     global.definition = require('../seeds/parserInitialGood.json')
 
@@ -68,7 +68,7 @@ describe('parser-body', () => {
     assert.deepStrictEqual(body, {"200":{"type":"array","items":{"type":"object","required":["name","id"],"properties":{"name":{"type":"string"},"tag":{"type":"string"},"id":{"type":"integer","format":"int64"}}}},"500":{"type":"object","required":["code","message"],"properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"}}}})
   })
 
-  it('items, ref and allof response openapi3.0', () => {
+  it('body items, ref and allof response openapi3.0', () => {
 
     global.definition = require('../seeds/parserInitialGoodOpenApiExpanded3.json')
 
@@ -76,15 +76,15 @@ describe('parser-body', () => {
     assert.deepStrictEqual(body, {"200":{"type":"array","items":{"type":"object","required":["name","id"],"properties":{"name":{"type":"string"},"tag":{"type":"string"},"id":{"type":"integer","format":"int64"}}}},"500":{"type":"object","required":["code","message"],"properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"}}}})
   })
 
-  it('items, ref and allof response openapi3.1', () => {
+  it('body items, ref and allof response openapi3.1', () => {
 
     global.definition = require('../seeds/parserInitialGoodOpenApiExpanded3.1.json')
 
-    const body = require('../src/parser/openapi31/body.js')('GET','/pets',true)
+    const body = require('../src/parser/openapi3/body.js')('GET','/pets',true)
     assert.deepStrictEqual(body, {"200":{"type":"array","items":{"type":"object","required":["name","id"],"properties":{"name":{"type":"string"},"tag":{"type":"string"},"id":{"type":"integer","format":"int64"}}}},"500":{"type":"object","required":["code","message"],"properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"}}}})
   })
 
-  it('ref and allof response swagger2', () => {
+  it('body ref and allof response swagger2', () => {
 
     global.definition = require('../seeds/parserInitialGood.json')
     global.circularTail = []
@@ -93,7 +93,7 @@ describe('parser-body', () => {
     assert.deepStrictEqual(body, {"200":{"type":"object","required":["name","id"],"properties":{"name":{"type":"string"},"tag":{"type":"string"},"id":{"type":"integer","format":"int64"}}},"500":{"type":"object","required":["code","message"],"properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"}}}})
   })
 
-  it('ref and allof response openapi3.0', () => {
+  it('body ref and allof response openapi3.0', () => {
 
     global.definition = require('../seeds/parserInitialGoodOpenApiExpanded3.json')
     global.circularTail = []
@@ -102,16 +102,16 @@ describe('parser-body', () => {
     assert.deepStrictEqual(body, {"200":{"type":"object","required":["name","id"],"properties":{"name":{"type":"string"},"tag":{"type":"string"},"id":{"type":"integer","format":"int64"}}},"500":{"type":"object","required":["code","message"],"properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"}}}})
   })
 
-  it('ref and allof response openapi3.1', () => {
+  it('body ref and allof response openapi3.1', () => {
 
     global.definition = require('../seeds/parserInitialGoodOpenApiExpanded3.1.json')
     global.circularTail = []
 
-    const body = require('../src/parser/openapi31/body.js')('POST','/pets',true)
+    const body = require('../src/parser/openapi3/body.js')('POST','/pets',true)
     assert.deepStrictEqual(body, {"200":{"type":"object","required":["name","id"],"properties":{"name":{"type":"string"},"tag":{"type":"string"},"id":{"type":"integer","format":"int64"}}},"500":{"type":"object","required":["code","message"],"properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"}}}})
   })
 
-  it('allof loop swagger2', () => {
+  it('body allof loop swagger2', () => {
 
     global.definition = require('../seeds/parserBodyAllofLoopInitial.json')
 
@@ -119,7 +119,7 @@ describe('parser-body', () => {
     assert.deepStrictEqual(body, {"required":["red","orange"],"properties":{"red":{"type":"string","maxLength":50},"orange":{"type":"number","maxLength":9},"water":{"required":["paramAiden"],"properties":{"paramAiden":{"type":"string"},"paramBiden":{"type":"string"},"paramCiden":{"type":"string"}},"type":"object"}},"type":"object"})
   })
 
-  it('allof string swagger2', () => {
+  it('body allof string swagger2', () => {
 
     global.definition = require('../seeds/parserBodyAllofString.json')
