+++ test/parser-queryParams.js
@@ -4,7 +4,7 @@ const assert = require('assert');
 
 describe('parser-queryParams', () => {
   
-  it('fill swagger2', () => {
+  it('queryParams fill swagger2', () => {
 
     global.definition = require('../seeds/parserQueryParamsGood.json')
 
@@ -21,7 +21,7 @@ describe('parser-queryParams', () => {
          } ])
   })
 
-  it('void swagger2', () => {
+  it('queryParams void swagger2', () => {
 
     global.definition = require('../seeds/parserQueryParamsGood.json')
 
@@ -29,7 +29,7 @@ describe('parser-queryParams', () => {
     assert.deepStrictEqual(queryParams, [])
   })
 
-  it('fill openapi3.0', () => {
+  it('queryParams fill openapi3.0', () => {
 
     global.definition = require('../seeds/parserInitialGoodOpenApiExpanded3.json')
 
@@ -46,7 +46,7 @@ describe('parser-queryParams', () => {
          } ])
   })
 
-  it('void openapi3.0', () => {
+  it('queryParams void openapi3.0', () => {
 
     global.definition = require('../seeds/parserInitialGoodOpenApiExpanded3.json')
 
@@ -54,11 +54,11 @@ describe('parser-queryParams', () => {
     assert.deepStrictEqual(queryParams, [])
   })
 
-  it('fill openapi3.1', () => {
+  it('queryParams fill openapi3.1', () => {
 
     global.definition = require('../seeds/parserInitialGoodOpenApiExpanded3.1.json')
 
-    const queryParams = require('../src/parser/openapi31/queryParams.js')('GET','/pets')
+    const queryParams = require('../src/parser/openapi3/queryParams.js')('GET','/pets')
     assert.deepStrictEqual(queryParams, [{
             "name": "tags",
             "required": false,
@@ -71,11 +71,11 @@ describe('parser-queryParams', () => {
          } ])
   })
 
-  it('void openapi3.1', () => {
+  it('queryParams void openapi3.1', () => {
 
     global.definition = require('../seeds/parserInitialGoodOpenApiExpanded3.1.json')
 
-    const queryParams = require('../src/parser/openapi31/queryParams.js')('POST','/pets')
+    const queryParams = require('../src/parser/openapi3/queryParams.js')('POST','/pets')
     assert.deepStrictEqual(queryParams, [])
   })
 })
